requirements.txt
beautifulsoup4==4.13.4
numpy==1.24.3
requests==2.32.5
python-dotenv==1.0.1
